requirements.txt
langchain==0.0.251
openai==0.27.8
python-dotenv==1.0.0
Flask==2.3.2
pypdf==3.15.0
tiktoken==0.4.0
chromadb==0.4.5
Flask-Cors==4.0.0
pdfminer.six==20221105
